Did some work on Listing & my listings pages
I added in a base.html with a simple navbar that currently just has raw links to all pages on the site. To make this navbar site wide ever html file should look like this

start
{% extends "base.html" %}
{% block content %}
Everything specific to this page
{% endblock %}
end

In the base I added in the CDN links for the bootstrap stylesheet as well as for the scripts for ajax and jquery because we may want to use these for updating pages and submitting forms.

Listing & my listing pages still need to be formatted better and need to display images.